// solteca/src/app/models/vendedor/recientes.ts
export interface Recientes {
  Id_carrito: number;
  Nombre_cliente: string;
  Origen: string;
  Destino: string;
  Tipo: string;
  Cantidad: number;
  Escala: string;
  Precio: number;
  Fecha_salida: string;
  Hora_salida: string;
  Telefono: string;
  Asiento: number;
  Referencia: string;
  Id_venta: string;
  Estado: number;
  Id_autobus: number;
  Id_sucursal: string;
  Numero_autobus: number;
  Status: number;
}